+++ Haden.NxtSDK/NxtExceptions.cs
@@ -1,0 +1,25 @@
+using System;
+
+namespace Haden.NxtSDK
+{
+    public class NxtProtocolException : ApplicationException
+    {
+        public NxtProtocolException(string message) : base(message)
+        {
+        }
+    }
+
+    public class NxtChannelBusyException : NxtProtocolException
+    {
+        public NxtChannelBusyException(string message) : base(message)
+        {
+        }
+    }
+
+    public class NxtCommunicationBusErrorException : NxtProtocolException
+    {
+        public NxtCommunicationBusErrorException(string message) : base(message)
+        {
+        }
+    }
+}